mt5-redis-publisher/requirements.txt
MetaTrader5>=5.0.5388
redis>=5.0.1
python-dotenv>=1.0.0
